rename source to streampredictor
