+++ streampredictor/Generalizer.py
