+++ streampredictor/GenerateText.py
