+++ streampredictor/ProtobufManager.py
