+++ streampredictor/scratch/StreamCounter.py
